+++ Konsole/IConsole.cs
@@ -9,6 +9,7 @@ namespace Konsole
         int CursorTop { get; set; }
         int CursorLeft { get; set; }
         ConsoleColor ForegroundColor { get; set; }
+        ConsoleColor BackgroundColor { get; set; }
         void SetCursorPosition(int x, int y);
         void PrintAt(int x, int y, string format, params object[] args);
         void PrintAt(int x, int y, string text);